backend/requirement.txt
flask
flask-cors
pyjwt
torch
librosa
tensorflow
transformers
numpy
paho-mqtt
unidecode
